Initial work on pty mode with terminal spawning

+++ src/char/char_loopback.c
@@ -99,7 +99,7 @@ typedef struct {
     int type;
     uint8_t data_tx;
     uint8_t data_rx;
-    uint8_t data_read : 1;
+    int data_read : 1;
     uint32_t control;
     uint32_t status;
 } char_loopback_t;
